Update p1.c
Line 4(apply to line 35 as well): In different projects there is a different standard of documenting the functions. I would propose to start practicing with the doxygen documentation:

/**
 * @brief      De_pus_scurta_descriere_a_functiei
 * @param[in]  fiecare_parametru_de_intrare_linie_separata_de_ex_aici_am_pune_doar_x : descriere_parametrului
 * @return       void
 */

Line 10: please use a relevant variable name(see finding for line 37)

Line 16: Functionality is not as expected.Please recheck and re-write the loop so that the output is the one expected.

Output:

Give the value for which to print the bits: 32

This is the number you gave written in binary representation:                                                                                                                      
00000100000000000000000000000000

Expected:

00000000000000000000000000100000   

Extra for Line 16: Do we really need to loop from all those values ? most of the times the values looped are not relevant for us and just add complexity.
E.g 4 => 0000000000000000000000000000100 ==>> only 3rd bit is relevant for the check here
      5 => 0000000000000000000000000000101 ==>> the same test is being performed again so this iteration is redundant
Can you think of a way of optimizing it ?

Line 25: The value of x is changed. In some situations further processing might be needed over the input value. Can you think of a way of implementing a solution without changing the input value ?

Line 18: // is more of a C++ style comments. I would suggest using /* */ type of comment even for single lines

Line 6,7,32,18: Sometimes too many comments are not relevant :)

Line 37: Always initialize the variables
             Always consider using a descriptive name for the variable. What is the objective of the variable x ? What is it's purpose ? Could you rename it to be more obvious for the reader ?

Line 39: Always check the official documentation when using standard functions . In this case please check https://www.man7.org/linux/man-pages/man3/scanf.3.html.
What will happen in case of errors ? How could this be handled ? E.g. : behavior when the input is negative or greater than uint32